+++ crates/test-ui/ui/extern_class_wrong_mutability.rs
@@ -19,40 +19,4 @@ extern_class!(
     }
 );
 
-extern_class!(
-    pub struct MyImmutableClass1;
-
-    unsafe impl ClassType for MyImmutableClass1 {
-        type Super = NSObject;
-        type Mutability = mutability::InteriorMutableWithSubclass<MyMutableClass1>;
-    }
-);
-
-extern_class!(
-    pub struct MyMutableClass1;
-
-    unsafe impl ClassType for MyMutableClass1 {
-        type Super = MyImmutableClass1;
-        type Mutability = mutability::InteriorMutableWithSuperclass<NSObject>;
-    }
-);
-
-extern_class!(
-    pub struct MyImmutableClass2;
-
-    unsafe impl ClassType for MyImmutableClass2 {
-        type Super = NSObject;
-        type Mutability = mutability::InteriorMutableWithSubclass<NSObject>;
-    }
-);
-
-extern_class!(
-    pub struct MyMutableClass2;
-
-    unsafe impl ClassType for MyMutableClass2 {
-        type Super = MyImmutableClass2;
-        type Mutability = mutability::InteriorMutableWithSuperclass<MyImmutableClass2>;
-    }
-);
-
 fn main() {}